#279: refactor ActiveYamlBase Currency.coins into AR model
